Move WorkerApiKtLintRunner into worker subpackage.
